test(omo-local-update): extract shared test helpers for multi-file coverage
