+++ packages/coding-agent/test/omo-local-update.test.ts
@@ -1,21 +1,17 @@
 import { execFileSync, spawn, spawnSync } from "node:child_process";
 import { createHash } from "node:crypto";
 import {
-	chmodSync,
 	cpSync,
 	existsSync,
 	mkdirSync,
-	mkdtempSync,
 	readdirSync,
 	readFileSync,
 	realpathSync,
 	renameSync,
 	rmSync,
-	statSync,
 	writeFileSync,
 } from "node:fs";
-import { tmpdir } from "node:os";
-import { delimiter, dirname, join, relative, resolve } from "node:path";
+import { dirname, join, relative, resolve } from "node:path";
 import { afterAll, describe, expect, it } from "vitest";
 import {
 	computeRemoteState,
@@ -41,37 +37,26 @@ import {
 	dirtyUntracked,
 	FIXTURE_PLUGIN_ARTIFACTS,
 } from "./omo-local-update-fixture.ts";
+import {
+	applyOmoGitIsolation,
+	artifactMtimes,
+	createTempRoots,
+	git,
+	installFakeBun,
+	makeAgentDir,
+	makeLogCollector,
+	makeSpyRun,
+	withPrependedPath,
+} from "./omo-local-update-helpers.ts";
 
-// Git determinism: isolate every git invocation in this file (test-side AND
-// engine-side through the inherited process.env) from ambient host config.
-const gitConfigDir = mkdtempSync(join(tmpdir(), "omo-local-update-gitcfg-"));
-const emptyGitConfig = join(gitConfigDir, "config");
-writeFileSync(emptyGitConfig, "");
-process.env.GIT_CONFIG_NOSYSTEM = "1";
-process.env.GIT_CONFIG_GLOBAL = emptyGitConfig;
-process.env.GIT_AUTHOR_NAME = "senpi-test";
-process.env.GIT_AUTHOR_EMAIL = "senpi-test@example.com";
-process.env.GIT_COMMITTER_NAME = "senpi-test";
-process.env.GIT_COMMITTER_EMAIL = "senpi-test@example.com";
-
-const tempRoots: string[] = [];
-
-function makeTempRoot(): string {
-	const root = mkdtempSync(join(tmpdir(), "omo-local-update-test-"));
-	tempRoots.push(root);
-	return root;
-}
+const gitIsolation = applyOmoGitIsolation();
+const tempRoots = createTempRoots();
+const makeTempRoot = tempRoots.makeTempRoot;
 
 afterAll(() => {
-	for (const root of tempRoots) {
-		rmSync(root, { recursive: true, force: true });
-	}
-	rmSync(gitConfigDir, { recursive: true, force: true });
+	tempRoots.cleanup();
+	gitIsolation.cleanup();
 });
-
-function git(args: string[], cwd: string): string {
-	return execFileSync("git", args, { cwd, encoding: "utf-8" }).trim();
-}
 
 interface LayoutOptions {
 	pluginPkgName?: string;
@@ -146,27 +131,6 @@ function currentBranch(cwd: string): string | undefined {
 	}
 }
 
-function makeLogCollector(): { lines: string[]; log: (message: string) => void } {
-	const lines: string[] = [];
-	return {
-		lines,
-		log: (message: string) => {
-			lines.push(message);
-		},
-	};
-}
-
-function makeSpyRun(): { calls: string[][]; run: OmoLocalRun } {
-	const calls: string[][] = [];
-	return {
-		calls,
-		run: (command, args, options) => {
-			calls.push([command, ...args]);
-			return defaultRun(command, args, options);
-		},
-	};
-}
-
 async function captureFailure(promise: Promise<unknown>): Promise<unknown> {
 	try {
 		await promise;
@@ -194,65 +158,6 @@ function snapshotDir(dir: string): Record<string, string> {
 		}
 	}
 	return snapshot;
-}
-
-function artifactMtimes(pluginPath: string): Record<string, number> {
-	const mtimes: Record<string, number> = {};
-	for (const artifact of FIXTURE_PLUGIN_ARTIFACTS) {
-		mtimes[artifact] = statSync(join(pluginPath, artifact)).mtimeMs;
-	}
-	return mtimes;
-}
-
-/**
- * Executable `bun` stand-in for orchestrator tests: `bun install` succeeds quietly;
- * `bun run build:senpi-plugin` runs the fixture's real stub build from the spawn cwd
- * (the updater's build worktree). FAKE_BUN_BUILD_FAIL makes the build exit 1 after
- * the stub wrote its artifacts into the worktree.
- */
-function installFakeBun(binDir: string): void {
-	mkdirSync(binDir, { recursive: true });
-	const script = [
-		"#!/bin/sh",
-		'if [ "$1" = "install" ]; then',
-		'  echo "fake bun install: ok"',
-		"  exit 0",
-		"fi",
-		'if [ "$1" = "run" ] && [ "$2" = "build:senpi-plugin" ]; then',
-		"  node scripts/build-senpi-plugin.mjs",
-		"  build_status=$?",
-		'  if [ -n "$FAKE_BUN_BUILD_FAIL" ]; then',
-		'    echo "fake bun: simulated build failure" >&2',
-		"    exit 1",
-		"  fi",
-		"  exit $build_status",
-		"fi",
-		'echo "fake bun: unexpected argv: $*" >&2',
-		"exit 1",
-		"",
-	].join("\n");
-	const bunPath = join(binDir, "bun");
-	writeFileSync(bunPath, script);
-	chmodSync(bunPath, 0o755);
-}
-
-/** Prepend binDir to PATH; returns a restore function. */
-function withPrependedPath(binDir: string): () => void {
-	const originalPath = process.env.PATH;
-	process.env.PATH = `${binDir}${delimiter}${originalPath ?? ""}`;
-	return () => {
-		if (originalPath === undefined) {
-			delete process.env.PATH;
-		} else {
-			process.env.PATH = originalPath;
-		}
-	};
-}
-
-function makeAgentDir(root: string): string {
-	const agentDir = join(root, "agent");
-	mkdirSync(agentDir, { recursive: true });
-	return agentDir;
 }
 
 describe("isKillSwitched", () => {
